% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{config_formatter}
\alias{config_formatter}
\title{格式化配置文件}
\usage{
config_formatter(data)
}
\arguments{
\item{data}{数据框的配置文件}
}
\value{
返回值
}
\description{
格式化配置文件
}
\examples{
config_formatter()
# config <- list(
#           row = list(
#             Fshow =TRUE,
#             Fname='第一行',
#             Fid='row',
#             Ficon='clock'
#           ),
#           column =list(
#             Fshow=FALSE,
#             Fname='测试一下',
#             Fid='column',
#             Ficon='history'
#           ),
#           book = list(
#             Fshow=TRUE,
#             Fname='第三项',
#             Fid='book',
#             Ficon='plane'
#           ),
#           series=list(
#             Fshow=TRUE,
#             Fname='第四项',
#             Fid='series',
#             Ficon='car-side'
#           ),
#           majority=list(
#             Fshow=FALSE,
#             Fname='第五项',
#             Fid='majority',
#             Ficon='hourglass'
#           
#           ),
#           tutor=list(
#             Fshow=TRUE,
#             Fname='第六项',
#             Fid='tutor',
#             Ficon='user-times'
#           )
#            
# )
data_config <- data.frame(Fshow=c(TRUE,TRUE,TRUE,TRUE,TRUE,TRUE),
Fname=LETTERS[1:6],
Fid = c('row','column','book','series','majority','tutor'),
Ficon=c('clock','history','plane','car-side','hourglass','user-times'),
stringsAsFactors = F)

格式化为列表

config <- config_formatter(data_config)

}
