% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{counter}
\alias{counter}
\title{处理计数器的逻辑}
\usage{
counter(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理计数器的逻辑
}
\examples{
counter();
}
