% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{counterButton}
\alias{counterButton}
\title{处理单个模块的问题}
\usage{
counterButton(id, label = "Counter")
}
\arguments{
\item{id}{内码}

\item{label}{标签}
}
\value{
返回值
}
\description{
处理单个模块的问题
}
\examples{
counterButton();
}
