% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{csvFile}
\alias{csvFile}
\title{处理服务器端的逻辑}
\usage{
csvFile(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{stringsAsFactors}{是否字符串}
}
\value{
返回值
}
\description{
处理服务器端的逻辑
}
\examples{
csvFile();
}
