% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{csvFileInput}
\alias{csvFileInput}
\title{文件输入器}
\usage{
csvFileInput(id, label = "CSV file")
}
\arguments{
\item{id}{命名空间内码}

\item{label}{标签}
}
\value{
返回值
}
\description{
文件输入器
}
\examples{
csvFileInput();
}
