% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{getUsers}
\alias{getUsers}
\title{获取用户登录信息}
\usage{
getUsers(conn = conn_rds("rdbe"), app_id = "appTpl")
}
\arguments{
\item{conn}{数据库信息}

\item{app_id}{指定APP}
}
\value{
返回值
}
\description{
获取用户登录信息
}
\examples{
getUsers()
}
