% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{inner}
\alias{inner}
\title{处理最里面的逻辑}
\usage{
inner(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理最里面的逻辑
}
\examples{
inner();
}
