% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.button.action.R
\docType{class}
\name{input.button.action-class}
\alias{input.button.action-class}
\title{定义input.button.action输入控件}
\value{
返回值
}
\description{
定义input.button.action输入控件
}
\examples{
input.button.download()
}
