% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.button.action.R
\name{input.button.action}
\alias{input.button.action}
\title{定义input.button.download的同名的入口函数}
\usage{
input.button.action(Id = "btn_action", label = "点击按纽", width = "100\%")
}
\arguments{
\item{Id}{内码}

\item{label}{标签}
}
\value{
返回值
}
\description{
定义input.button.download的同名的入口函数
}
\examples{
input.button.action();
}
