% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{linkedScatter}
\alias{linkedScatter}
\title{处理相关的服务器}
\usage{
linkedScatter(input, output, session, data, left, right)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{data}{数据}

\item{left}{左}

\item{right}{右}
}
\value{
返回值
}
\description{
处理相关的服务器
}
\examples{
linkedScatter();
}
