% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_action_button}
\alias{mdl_action_button}
\title{处理点击按纽}
\usage{
mdl_action_button(id, label = "点击按纽", is.test = FALSE)
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{is.test}{是否测试}
}
\value{
返回值
}
\description{
处理点击按纽
}
\examples{
mdl_action_button();
}
