% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{mdl_dataTable}
\alias{mdl_dataTable}
\title{处理dataTable数据}
\usage{
mdl_dataTable(id, label = "dataTable")
}
\arguments{
\item{id}{内码}
}
\value{
返回值
}
\description{
处理dataTable数据
}
\examples{
mdl_dataTable();
}
