% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_date}
\alias{mdl_date}
\title{处理日期字段}
\usage{
mdl_date(id, label = "日期", is.test = FALSE, value = Sys.Date())
}
\arguments{
\item{id}{id}

\item{label}{标签}

\item{is.test}{是否测试}

\item{value}{默认值}
}
\value{
返回值
}
\description{
处理日期字段
}
\examples{
mdl_date();
library(tsui);
library(shiny);
test_tsui(mdl_date,id='date1','日期',run_date_test);
}
