% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_dateRange}
\alias{mdl_dateRange}
\title{处理日期范围字段}
\usage{
mdl_dateRange(id, label = "日期范围", is.test = FALSE,
  startDate = Sys.Date(), endDate = Sys.Date() + 14)
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{is.test}{是否测试}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
处理日期范围字段
}
\examples{
mdl_dateRange();
}
