% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{mdl_download_button}
\alias{mdl_download_button}
\title{测试下载按钮}
\usage{
mdl_download_button(id, label)
}
\arguments{
\item{id}{内码}

\item{label}{标签}
}
\value{
返回值
}
\description{
测试下载按钮
}
\examples{
mdl_download_button();
}
