% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_file}
\alias{mdl_file}
\title{处理单一文件控制}
\usage{
mdl_file(id, label = "单文件选择器：", is.test = FALSE,
  fileType = ".xlsx")
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{is.test}{是否测试}

\item{fileType}{文件后缀类型xlsx}
}
\value{
返回值
}
\description{
处理单一文件控制
}
\examples{
mdl_file();
}
