% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_files}
\alias{mdl_files}
\title{多个文件上传处理器}
\usage{
mdl_files(id, label = "多文件选择器：", is.test = FALSE,
  fileType = ".xlsx")
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{is.test}{是否测试}

\item{fileType}{文件类型}
}
\value{
返回值
}
\description{
多个文件上传处理器
}
\examples{
mdl_files();
}
