% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_integer}
\alias{mdl_integer}
\title{处理整数控件}
\usage{
mdl_integer(
  id,
  label = "整数",
  is.test = FALSE,
  min = 0,
  max = 10,
  value = 3,
  step = 1,
  round = 0L,
  width = "100\%"
)
}
\arguments{
\item{id}{id}

\item{label}{标签}

\item{is.test}{是否测试}

\item{min}{最小值}

\item{max}{最大值}

\item{value}{默认值}

\item{step}{步长1}

\item{round}{小数位}

\item{width}{宽度}
}
\value{
返回值
}
\description{
处理整数控件
}
\examples{
library(shiny);
library(tsui);
test_tsui(mdl_integer,id='int1','整数',run_integer_test);
}
