% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_businessUI.R
\name{mdl_ListChoose1}
\alias{mdl_ListChoose1}
\title{从单项选择题中选择元素}
\usage{
mdl_ListChoose1(id, label = "单项选择题",
  choiceNames = list(icon("calendar"), icon("bed"), icon("cog"),
  icon("bug")), choiceValues = list("calendar", "bed", "cog", "bug"),
  selected = "bed")
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{choiceNames}{显示选项}

\item{choiceValues}{存储选项}

\item{selected}{是否选中}
}
\value{
返回值
}
\description{
从单项选择题中选择元素
}
\examples{
mdl_ListChoose1();
}
