% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_logical}
\alias{mdl_logical}
\title{处理逻辑值}
\usage{
mdl_logical(id, label, value = FALSE)
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{value}{默认值FALSE}
}
\value{
返回值
}
\description{
处理逻辑值
}
\examples{
mdl_logical();
}
