% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_rdbms.R
\name{mdl_mysql_conn}
\alias{mdl_mysql_conn}
\title{设置mysql的连接信息}
\usage{
mdl_mysql_conn(id, label = "MySQL配置信息")
}
\arguments{
\item{id}{内码}

\item{label}{标签}
}
\value{
返回值
}
\description{
设置mysql的连接信息
}
\examples{
mdl_mysql_conn();
}
