% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_ns.R
\name{mdl_ns_brand}
\alias{mdl_ns_brand}
\title{处理网商品牌订单界面信息}
\usage{
mdl_ns_brand(id, label)
}
\arguments{
\item{id}{内码信息}

\item{label}{标题信息}
}
\value{
返回值
}
\description{
处理网商品牌订单界面信息
}
\examples{
mdl_ns_brand('ns_select','请选择品牌');
}
