% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_password}
\alias{mdl_password}
\title{处理密码控件}
\usage{
mdl_password(id, label = "密码", is.test = FALSE)
}
\arguments{
\item{id}{密码ID}

\item{label}{标签名称}

\item{is.test}{是否测试}
}
\value{
返回值
}
\description{
处理密码控件
}
\examples{
library(shiny);
library(tsui);
test_tsui(mdl_password,'123',label = '密码:',f_test = run_password_test);
}
