% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_brand.R
\name{mdl_question}
\alias{mdl_question}
\title{处理问题列表}
\usage{
mdl_question(id, data, button_txt = "提交答案")
}
\arguments{
\item{id}{内码}

\item{data}{数据}

\item{button_txt}{提交答案}
}
\value{
返回值
}
\description{
处理问题列表
}
\examples{
mdl_question();
}
