% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_rdbms.R
\name{mdl_sql_conn}
\alias{mdl_sql_conn}
\title{撰写数据库连接组件}
\usage{
mdl_sql_conn(id, label = "SQL Server配置信息",
  ip = "115.159.201.178", port = "1433", user_name = "sa",
  password = "88888888", db_name = "demo")
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{ip}{地址}

\item{port}{端口}

\item{user_name}{用户名}

\item{password}{密码}

\item{db_name}{数据库名称}
}
\value{
返回值
}
\description{
撰写数据库连接组件
}
\examples{
library(shiny)
library(tsui);


ui <- fluidPage(
   
   
   titlePanel("SQL SERVER 数据库连接信息测试"),
   
   
   sidebarLayout(
      sidebarPanel(
         
         mdl_sql_conn('sql')
         
      ),
      
      mainPanel(
         mdl_dataTable('select')
      )
   )
)


server <- function(input, output) {
   
   con <- run_sql_common('sql');
   
   data <- reactive(
      {
         res  <- RJDBC::dbGetQuery(con(),'select top 100 * from T_GL_VOUCHER;')
         return(res)
      }
      
   )
   run_dataTable(proc_dataTable,id = 'select',data=data())
   
   
}

# Run the application 
shinyApp(ui = ui, server = server)
}
