% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_advancedUI.R
\name{mdl_sql_query}
\alias{mdl_sql_query}
\title{处理查询组件}
\usage{
mdl_sql_query(
  id,
  label = "SQL查询分析器：",
  value = c("select * \\n from t_item;"),
  fontSize = 16,
  theme = 10,
  height = "300px"
)
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{value}{默认值}

\item{fontSize}{字体大小}

\item{theme}{主题}

\item{height}{高度}
}
\value{
返回值
}
\description{
处理查询组件
}
\examples{
mdl_sql_query();
}
