% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{mdl_text}
\alias{mdl_text}
\title{编写模块化的文本字段}
\usage{
mdl_text(id, label = "文本", is.test = FALSE)
}
\arguments{
\item{id}{内码字段}

\item{label}{标签字段}

\item{is.test}{是否测试模式}
}
\value{
返回值
}
\description{
编写模块化的文本字段
}
\examples{
示例1 shiny应用程序: 
library(shiny);
library(tsui);
ui <- fluidPage(
 mdl_text('text1',label = '姓名',is.test = T),
 mdl_text('text2',label = '密码',is.test = T)
)

server <- function(input, output, session) {
 run_text_test('text1')
 run_text_test('text2')
 
}

shinyApp(ui, server)
}
