% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_text2.R
\name{mdl_text2}
\alias{mdl_text2}
\title{处理数据}
\usage{
mdl_text2(id, label = "文本", value = "", is.test = FALSE)
}
\arguments{
\item{id}{内码}

\item{label}{文件}

\item{value}{值}

\item{is.test}{是否测试}
}
\value{
返回值
}
\description{
处理数据
}
\examples{
mdl_text2
}
