% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_advancedUI.R
\name{mdl_textEditor}
\alias{mdl_textEditor}
\title{处理编辑器模块}
\usage{
mdl_textEditor(
  id,
  label = "多行文本编码器输入：",
  is.test = FALSE,
  value = c("文本编码器测试\\n测试换行"),
  fontSize = 16,
  theme = 10,
  height = "200px"
)
}
\arguments{
\item{id}{内码}

\item{label}{标签}

\item{is.test}{是否测试}

\item{value}{默认值}

\item{fontSize}{字体大小}

\item{theme}{主题编号}

\item{height}{默认编码器高级}
}
\value{
返回值
}
\description{
处理编辑器模块
}
\examples{
mdl_textEditor();
}
