% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{outer}
\alias{outer}
\title{处理外面的逻辑}
\usage{
outer(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理外面的逻辑
}
\examples{
outer();
}
