% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_dashboard_default.R
\name{page_dashboard_default}
\alias{page_dashboard_default}
\title{定义仪表盘风格的页面}
\usage{
page_dashboard_default(HeadingName = "Basic dashboard", leftAreaList,
  rightAreaList)
}
\arguments{
\item{HeadingName}{标题栏}

\item{leftAreaList}{左内容}

\item{rightAreaList}{右内容}
}
\value{
返回值
}
\description{
定义仪表盘风格的页面
}
\examples{
page_dashboard_default();
}
