% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_fluid_4.R
\name{page_fluid_4}
\alias{page_fluid_4}
\title{定义page_fluid_田字形模板4个元素}
\usage{
page_fluid_4(HeadingName, leftAreaList, R1C1List, R1C2List, R2C1List,
  R2C2List)
}
\arguments{
\item{HeadingName}{标题}

\item{leftAreaList}{左区分内容}

\item{R1C1List}{第一行第一列取值内容}

\item{R1C2List}{第一行第二列取值内容}

\item{R2C1List}{第二行第一列取值内容}

\item{R2C2List}{第二行第二列取值内容}
}
\value{
返回值
}
\description{
定义page_fluid_田字形模板4个元素
}
\examples{
HeadingName <-"Old Faithful Geyser Data";
leftAreaList <- tagList( sliderInput("bins",
                                    "Number of bins:",
                                    min = 1,
                                    max = 50,
                                    value = 30),
                        fileInput('file1','文件'));
R1C1List <- tagList(fileInput('file1','文件R1C1'));
R2C1List <- tagList(fileInput('file2','文件R2C1'));
R1C2List <- tagList(fileInput('file3','文件R1C2'));
R2C2List <- tagList(fileInput('file4','文件R2C2'));


ui <- page_fluid_4(HeadingName,leftAreaList,R1C1List,R1C2List,R2C1List,R2C2List);
}
