% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_fluid_default.R
\name{page_fluid_default}
\alias{page_fluid_default}
\title{定义page_fluid的默认模板}
\usage{
page_fluid_default(HeadingName, leftAreaList, rightAreaList)
}
\arguments{
\item{HeadingName}{标题}

\item{leftAreaList}{左区分内容}

\item{rightAreaList}{右区分内容}
}
\value{
返回值
}
\description{
定义page_fluid的默认模板
}
\examples{
HeadingName <-"Old Faithful Geyser Data";
leftAreaList <- tagList( sliderInput("bins",
                                    "Number of bins:",
                                    min = 1,
                                    max = 50,
                                    value = 30),
                        fileInput('file1','文件'));
rightAreaList <-tagList(plotOutput("distPlot"));


ui <- page_fluid_default(HeadingName,leftAreaList,rightAreaList);
}
