% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{password_checkOriginal}
\alias{password_checkOriginal}
\title{check the orginal password}
\usage{
password_checkOriginal(fappId = "appTpl", fuser, fpassword)
}
\arguments{
\item{fappId}{app}

\item{fuser}{user}

\item{fpassword}{pwd}
}
\value{
return value
}
\description{
check the orginal password
}
\examples{
password_checkOriginal()
}
