% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{password_equal}
\alias{password_equal}
\title{对比两个密码}
\usage{
password_equal(newPassword, repPassword)
}
\arguments{
\item{newPassword}{新密码}

\item{repPassword}{重复密码}
}
\value{
返回是否相等
}
\description{
对比两个密码
}
\examples{
password_equal()
}
