% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{password_setNew}
\alias{password_setNew}
\title{数据库设置新密码}
\usage{
password_setNew(fappId = "appTpl", fuser, fpassword)
}
\arguments{
\item{fappId}{程序}

\item{fuser}{用户}

\item{fpassword}{密码}
}
\value{
返回值
}
\description{
数据库设置新密码
}
\examples{
password_setNew()
}
