% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_action_button}
\alias{proc_action_button}
\title{处理点击按纽}
\usage{
proc_action_button(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理点击按纽
}
\examples{
proc_action_button();
}
