% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_action_button_var}
\alias{proc_action_button_var}
\title{处理按纽处理逻辑}
\usage{
proc_action_button_var(input, output, session, init_null = FALSE)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{init_null}{初始化值是事为null}
}
\value{
返回值
}
\description{
处理按纽处理逻辑
}
\examples{
proc_action_button_var();
}
