% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_brand.R
\name{proc_brand_tpl}
\alias{proc_brand_tpl}
\title{提供处理逻辑}
\usage{
proc_brand_tpl(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
提供处理逻辑
}
\examples{
proc_brand_tpl();
}
