% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{proc_dataTable}
\alias{proc_dataTable}
\title{处理dataTable数据}
\usage{
proc_dataTable(input, output, session, data = iris)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{data}{数据}
}
\value{
返回值
}
\description{
处理dataTable数据
}
\examples{
proc_dataTable();
}
