% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_date}
\alias{proc_date}
\title{处理日期的逻辑框架}
\usage{
proc_date(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理日期的逻辑框架
}
\examples{
proc_date();
}
