% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{proc_download_button}
\alias{proc_download_button}
\title{编写处理逻辑的框架}
\usage{
proc_download_button(input, output, session, data = iris,
  filename = "下载文件.xlsx", func = write.xlsx)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{data}{需要下载的数据R对象}

\item{filename}{文件名}

\item{func}{处理函数}
}
\value{
返回值
}
\description{
编写处理逻辑的框架
}
\examples{
proc_download_button();
}
