% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_file}
\alias{proc_file}
\title{处理单个文件逻辑}
\usage{
proc_file(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理单个文件逻辑
}
\examples{
proc_file();
}
