% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_file_var}
\alias{proc_file_var}
\title{处理文件变量逻辑}
\usage{
proc_file_var(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理文件变量逻辑
}
\examples{
proc_file_var();
}
