% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_files}
\alias{proc_files}
\title{撰写逻辑处理通用框架}
\usage{
proc_files(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
撰写逻辑处理通用框架
}
\examples{
proc_files();
}
