% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_files_var}
\alias{proc_files_var}
\title{批量处理多个文件的逻辑}
\usage{
proc_files_var(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
批量处理多个文件的逻辑
}
\examples{
proc_files_var();
}
