% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_integer_var}
\alias{proc_integer_var}
\title{处理整数的逻辑函数}
\usage{
proc_integer_var(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理整数的逻辑函数
}
\examples{
proc_integer_var();
}
