% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_businessUI.R
\name{proc_ListChoose1}
\alias{proc_ListChoose1}
\title{处理多项选择的逻辑函数}
\usage{
proc_ListChoose1(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理多项选择的逻辑函数
}
\examples{
proc_ListChoose1();
}
