% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_logical}
\alias{proc_logical}
\title{处理逻辑数据}
\usage{
proc_logical(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理逻辑数据
}
\examples{
proc_logical();
}
