% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_login.R
\name{proc_login}
\alias{proc_login}
\title{处理登录窗口逻辑}
\usage{
proc_login(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理登录窗口逻辑
}
\examples{
proc_login();
}
