% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_rdbms.R
\name{proc_mysql_common}
\alias{proc_mysql_common}
\title{处理mysql的逻辑}
\usage{
proc_mysql_common(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理mysql的逻辑
}
\examples{
proc_mysql_common();
}
