% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_ns.R
\name{proc_ns_brand}
\alias{proc_ns_brand}
\title{处理品牌的业务处理逻辑}
\usage{
proc_ns_brand(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理品牌的业务处理逻辑
}
\examples{
proc_ns_brand();
}
