% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_numeric}
\alias{proc_numeric}
\title{定义函数的处理逻辑}
\usage{
proc_numeric(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义函数的处理逻辑
}
\examples{
proc_numeric();
}
