% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_numeric_var}
\alias{proc_numeric_var}
\title{获取变量的处理逻辑}
\usage{
proc_numeric_var(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
获取变量的处理逻辑
}
\examples{
proc_numeric_var();
}
