% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_password}
\alias{proc_password}
\title{定义处理密码字段的控件}
\usage{
proc_password(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义处理密码字段的控件
}
\examples{
proc_password();
}
