% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_password_var}
\alias{proc_password_var}
\title{处理文本的变量处理方式}
\usage{
proc_password_var(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理文本的变量处理方式
}
\examples{
proc_password_var();
}
