% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{proc_print}
\alias{proc_print}
\title{使用print的处理逻辑
#这样的处理逻辑有一个问题，当数据变化时函数并不知晓。}
\usage{
proc_print(input, output, session, data)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{data}{数据}
}
\value{
返回值
}
\description{
使用print的处理逻辑
#这样的处理逻辑有一个问题，当数据变化时函数并不知晓。
}
\examples{
proc_print();
}
