% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_brand.R
\name{proc_question}
\alias{proc_question}
\title{处理标准问的逻辑}
\usage{
proc_question(input, output, session, table_name = "ques_standard_answer")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{table_name}{写入表名}
}
\value{
返回值
}
\description{
处理标准问的逻辑
}
\examples{
proc_question();
}
