% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_rdbms.R
\name{proc_sql_common}
\alias{proc_sql_common}
\title{定义数据库连接的处理信息}
\usage{
proc_sql_common(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义数据库连接的处理信息
}
\examples{
proc_sql_common();
}
