% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_advancedUI.R
\name{proc_sql_query}
\alias{proc_sql_query}
\title{定义查询分析器的逻辑处理单元}
\usage{
proc_sql_query(input, output, session, conn)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{conn}{连接}
}
\value{
返回值
}
\description{
定义查询分析器的逻辑处理单元
}
\examples{
proc_sql_query();
}
