% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{proc_text}
\alias{proc_text}
\title{定义通过的text处理模板}
\usage{
proc_text(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义通过的text处理模板
}
\examples{
proc_text();
}
