% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_advancedUI.R
\name{proc_textEditor}
\alias{proc_textEditor}
\title{处理文本编码器的计算逻辑}
\usage{
proc_textEditor(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理文本编码器的计算逻辑
}
\examples{
proc_textEditor();
}
