% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUser.R
\name{readUserFile}
\alias{readUserFile}
\title{读取用户文件}
\usage{
readUserFile(file = "data-raw/sample_users.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取用户文件
}
\examples{
readUserFile()
}
