% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_action_button}
\alias{run_action_button}
\title{处理按纽事件}
\usage{
run_action_button(proc_func, id)
}
\arguments{
\item{proc_func}{逻辑函数}

\item{id}{内码}
}
\value{
返回值
}
\description{
处理按纽事件
}
\examples{
run_action_button(); 
}
