% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_action_button_test}
\alias{run_action_button_test}
\title{处理相应的按纽处理事项}
\usage{
run_action_button_test(id)
}
\arguments{
\item{id}{内码}
}
\value{
返回值
}
\description{
处理相应的按纽处理事项
}
\examples{
run_action_button_test();
}
