% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_dataTable}
\alias{run_dataTable}
\title{运行dataTable数据}
\usage{
run_dataTable(proc_func, id, data = iris)
}
\arguments{
\item{proc_func}{逻辑函数}

\item{id}{内码}

\item{data}{数据}
}
\value{
返回值
}
\description{
运行dataTable数据
}
\examples{
run_dataTable();
}
