% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_dataTable2}
\alias{run_dataTable2}
\title{处理改进版的处理函数}
\usage{
run_dataTable2(id, data = iris)
}
\arguments{
\item{id}{内码}

\item{data}{数据}
}
\value{
返回值
}
\description{
处理改进版的处理函数
}
\examples{
run_dataTable2();
}
