% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_dataTable_test}
\alias{run_dataTable_test}
\title{测试dataTable的应用}
\usage{
run_dataTable_test(id)
}
\arguments{
\item{id}{内码}
}
\value{
返回值
}
\description{
测试dataTable的应用
}
\examples{
run_dataTable_test();
}
