% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_dateRange}
\alias{run_dateRange}
\title{处理日期字段}
\usage{
run_dateRange(proc_func, id)
}
\arguments{
\item{proc_func}{日期处理逻辑}

\item{id}{内码}
}
\value{
返回值
}
\description{
处理日期字段
}
\examples{
run_dateRange();
}
