% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_download_button}
\alias{run_download_button}
\title{处理相应的逻辑}
\usage{
run_download_button(
  proc_func,
  id,
  data = iris,
  filename = "下载文件.xlsx",
  func = write.xlsx
)
}
\arguments{
\item{proc_func}{相应的函数}

\item{id}{内码}

\item{data}{数据}

\item{filename}{文件名}

\item{func}{相关的处理函数名,必须有data与file作为入口参数}
}
\value{
返回值
}
\description{
处理相应的逻辑
}
\examples{
run_download_button(); 
}
