% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_download_xlsx}
\alias{run_download_xlsx}
\title{处理下载文件}
\usage{
run_download_xlsx(id, data = iris, filename = "下载文件.xlsx")
}
\arguments{
\item{id}{内码}

\item{data}{数据}

\item{filename}{文件名}
}
\value{
返回值
}
\description{
处理下载文件
}
\examples{
run_download_xlsx();
}
