% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_numeric}
\alias{run_numeric}
\title{定义相应的处理函数}
\usage{
run_numeric(proc_func, id)
}
\arguments{
\item{proc_func}{逻辑函数}

\item{id}{ID}
}
\value{
返回值
}
\description{
定义相应的处理函数
}
\examples{
run_numeric();
}
