% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_password}
\alias{run_password}
\title{定义处理密码的逻辑函数}
\usage{
run_password(proc_func, id)
}
\arguments{
\item{proc_func}{逻辑函数名}

\item{id}{inputID}
}
\value{
返回值
}
\description{
定义处理密码的逻辑函数
}
\examples{
run_password();
}
