% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_password_test}
\alias{run_password_test}
\title{定义处理密码的演示逻辑}
\usage{
run_password_test(id)
}
\arguments{
\item{id}{输入密码控件ID}
}
\value{
返回值
}
\description{
定义处理密码的演示逻辑
}
\examples{
run_password_test();
}
