% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{run_print}
\alias{run_print}
\title{运行打印函数}
\usage{
run_print(id, data)
}
\arguments{
\item{id}{内码}

\item{data}{数据}
}
\value{
返回值
}
\description{
运行打印函数
}
\examples{
data <- var_text('aa');

observe({
 data();
 run_print('res',data());  
})
}
