% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_brand.R
\name{run_question}
\alias{run_question}
\title{运行相应的代码}
\usage{
run_question(id, table_name = "ques_standard_answer")
}
\arguments{
\item{id}{内码}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
运行相应的代码
}
\examples{
run_question()
}
