% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_rdbms.R
\name{run_sql_common}
\alias{run_sql_common}
\title{定义数据库连接的处理逻辑}
\usage{
run_sql_common(id)
}
\arguments{
\item{id}{内码}
}
\value{
返回值
}
\description{
定义数据库连接的处理逻辑
}
\examples{
library(shiny)
library(tsui);


ui <- fluidPage(
   
   
   titlePanel("SQL SERVER 数据库连接信息测试"),
   
   
   sidebarLayout(
      sidebarPanel(
         
         mdl_sql_conn('sql')
         
      ),
      
      mainPanel(
         mdl_dataTable('select')
      )
   )
)


server <- function(input, output) {
   
   con <- run_sql_common('sql');
   
   data <- reactive(
      {
         res  <- RJDBC::dbGetQuery(con(),'select top 100 * from T_GL_VOUCHER;')
         return(res)
      }
      
   )
   run_dataTable(proc_dataTable,id = 'select',data=data())
   
   
}

# Run the application 
shinyApp(ui = ui, server = server)
}
