% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_advancedUI.R
\name{run_sql_query}
\alias{run_sql_query}
\title{定义查询分析器的处理单元}
\usage{
run_sql_query(id, conn)
}
\arguments{
\item{id}{内码}

\item{conn}{连接}
}
\value{
返回值
}
\description{
定义查询分析器的处理单元
}
\examples{
run_sql_query();
}
