% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{run_text}
\alias{run_text}
\title{指定text运行处理函数}
\usage{
run_text(proc_func, id)
}
\arguments{
\item{proc_func}{逻辑函数}

\item{id}{input对应的id}
}
\value{
返回值
}
\description{
指定text运行处理函数
}
\examples{
run_text();
}
