% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{scatterPlot}
\alias{scatterPlot}
\title{处理函数处理逻辑}
\usage{
scatterPlot(data, cols)
}
\arguments{
\item{data}{数据}

\item{cols}{位置}
}
\value{
返回值
}
\description{
处理函数处理逻辑
}
\examples{
scatterPlot();
}
