% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTemplateV2.R
\name{setSideBarMenu}
\alias{setSideBarMenu}
\title{根据角色与数据库设置边栏}
\usage{
setSideBarMenu(
  conn = conn_rds("rdbe"),
  app_id = "appTpl",
  permission = "admin"
)
}
\arguments{
\item{conn}{数据库连接}

\item{app_id}{程序ID}

\item{permission}{角色}
}
\value{
返回值
}
\description{
根据角色与数据库设置边栏
}
\examples{
setSideBarMenu()
}
