% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_action_button}
\alias{test_action_button}
\title{处理按纽}
\usage{
test_action_button(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理按纽
}
\examples{
test_action_button(); 
}
