% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseOutput.R
\name{test_dataTable}
\alias{test_dataTable}
\title{处理测试数据集}
\usage{
test_dataTable(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理测试数据集
}
\examples{
test_dataTable();
}
