% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_dateRange}
\alias{test_dateRange}
\title{处理日期字段}
\usage{
test_dateRange(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理日期字段
}
\examples{
test_dateRange();
}
