% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_file}
\alias{test_file}
\title{测试文件}
\usage{
test_file(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
测试文件
}
\examples{
test_file();
}
