% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_files}
\alias{test_files}
\title{测试多个文件的处理}
\usage{
test_files(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
测试多个文件的处理
}
\examples{
test_files();
}
