% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_integer}
\alias{test_integer}
\title{定义示例的整数处理逻辑}
\usage{
test_integer(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义示例的整数处理逻辑
}
\examples{
test_integer();
}
