% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_numeric}
\alias{test_numeric}
\title{定义numeric测试逻辑}
\usage{
test_numeric(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义numeric测试逻辑
}
\examples{
test_numeric();
}
