% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{test_text}
\alias{test_text}
\title{定义一个测试的text处理输入}
\usage{
test_text(input, output, session)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
定义一个测试的text处理输入
}
\examples{
test_text();
}
