% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{test_tsui}
\alias{test_tsui}
\title{定义test_tsui的测试函数}
\usage{
test_tsui(f, id, label, f_test)
}
\arguments{
\item{f}{input_func}

\item{id}{id}

\item{label}{标签}

\item{f_test}{逻辑test函数}
}
\value{
返回值
}
\description{
定义test_tsui的测试函数
}
\examples{
test_tsui();
}
