% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,input.button.action-method}
\alias{tsui,input.button.action-method}
\title{tsui用于input.button.acton控制上面}
\usage{
\S4method{tsui}{input.button.action}(x)
}
\arguments{
\item{x}{input.button.action. 实例化}
}
\value{
返回值
}
\description{
tsui用于input.button.acton控制上面
}
\examples{
tsui();
}
