% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,input.button.download-method}
\alias{tsui,input.button.download-method}
\title{用于生成下载按纽}
\usage{
\S4method{tsui}{input.button.download}(x)

\S4method{tsui}{input.button.download}(x)
}
\arguments{
\item{x}{input.button.download.}
}
\value{
返回值

返回值
}
\description{
用于生成下载按纽

Title
}
\examples{
drawElement(input.button.download());
tsui(input.button.download());
}
