% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,input.date-method}
\alias{tsui,input.date-method}
\title{定义tsui方法在input.date上的实现}
\usage{
\S4method{tsui}{input.date}(x)
}
\arguments{
\item{x}{input.date. 实例化对象}
}
\value{
返回值
}
\description{
定义tsui方法在input.date上的实现
}
\examples{
tsui();
}
