% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,input.dateRange-method}
\alias{tsui,input.dateRange-method}
\title{定义处理日期范围的控制}
\usage{
\S4method{tsui}{input.dateRange}(x)
}
\arguments{
\item{x}{input.dateRange. 日期范围}
}
\value{
返回值
}
\description{
定义处理日期范围的控制
}
\examples{
tsui();
}
