% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,input.file-method}
\alias{tsui,input.file-method}
\title{定义tsui方法在input.file上的实现}
\usage{
\S4method{tsui}{input.file}(x)
}
\arguments{
\item{x}{input.file.  input.file类}
}
\value{
返回值
}
\description{
定义tsui方法在input.file上的实现
}
\examples{
 tsui(input.file());
}
