% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,output.dataTable-method}
\alias{tsui,output.dataTable-method}
\title{定义tsui方法在output.dataTable类的实现}
\usage{
\S4method{tsui}{output.dataTable}(x)
}
\arguments{
\item{x}{output.dataTable. 类output.dataTable实例对象}
}
\value{
返回值
}
\description{
定义tsui方法在output.dataTable类的实现
}
\examples{
tsui(output.dataTable());
}
