% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,output.image-method}
\alias{tsui,output.image-method}
\title{将于处理相应的数据}
\usage{
\S4method{tsui}{output.image}(x)
}
\arguments{
\item{x}{output.image.  output.image类的对象}
}
\value{
返回值
}
\description{
将于处理相应的数据
}
\examples{
drawElement();
}
