% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,output.plot-method}
\alias{tsui,output.plot-method}
\title{Title}
\usage{
\S4method{tsui}{output.plot}(x)
}
\arguments{
\item{x}{output.plot.  output.plot类的对象}
}
\value{
返回值
}
\description{
Title
}
\examples{
drawElement();
}
