% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,output.text-method}
\alias{tsui,output.text-method}
\title{定义drawElement方法在output.text上的实现}
\usage{
\S4method{tsui}{output.text}(x)
}
\arguments{
\item{x}{output.text.定义output.text的实例}
}
\value{
返回值
}
\description{
定义drawElement方法在output.text上的实现
}
\examples{
drawElement();
}
