% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,ui_date-method}
\alias{tsui,ui_date-method}
\title{shiny_input for ui_date}
\usage{
\S4method{tsui}{ui_date}(x)
}
\arguments{
\item{x}{ui_date.  对象}
}
\value{
返回值
}
\description{
shiny_input for ui_date
}
\examples{
shiny_input(x);
}
