% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,ui_file-method}
\alias{tsui,ui_file-method}
\title{定义shiny_input在ui_file上的实现}
\usage{
\S4method{tsui}{ui_file}(x)
}
\arguments{
\item{x}{ui_file.  ui_file实例对象}
}
\value{
返回值
}
\description{
定义shiny_input在ui_file上的实现
}
\examples{
shinput(ui_file_ins);
}
