% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,ui_password-method}
\alias{tsui,ui_password-method}
\title{定义shiny_input对ui_password的实现}
\usage{
\S4method{tsui}{ui_password}(x)
}
\arguments{
\item{x}{ui_password. 实例化对象ui_password}
}
\value{
返回值
}
\description{
定义shiny_input对ui_password的实现
}
\examples{
tsui(x);
}
