% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_defineClasses.R
\docType{class}
\name{ui_text-class}
\alias{ui_text-class}
\title{定义文件类}
\value{
返回值
}
\description{
定义文件类
}
\section{Slots}{

\describe{
\item{\code{inputId}}{character.}

\item{\code{label}}{character.}

\item{\code{value}}{character.}

\item{\code{width}}{character.}
}}

\examples{
aa
}
