% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUser.R
\name{userInfo_upload}
\alias{userInfo_upload}
\title{用户详细描述上传}
\usage{
userInfo_upload(conn = conn_rds("rdbe"), data, app_id)
}
\arguments{
\item{conn}{连接}

\item{data}{数据}

\item{app_id}{程序id}
}
\value{
返回值
}
\description{
用户详细描述上传
}
\examples{
userInfo_upload()
}
