% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUser.R
\name{userQueryField}
\alias{userQueryField}
\title{查询用户字段}
\usage{
userQueryField(conn = conn_rds("rdbe"), app_id = "appTpl",
  user = "demo", key = "Fname")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{user}{用户}

\item{key}{关键字}
}
\value{
返回值
}
\description{
查询用户字段
}
\examples{
userQueryField()
}
