% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUser.R
\name{userRight_upload}
\alias{userRight_upload}
\title{上传用户数据}
\usage{
userRight_upload(
  conn = conn_rds("rdbe"),
  app_id = "appTpl",
  data = readUserFile(),
  default_pwd = "123456"
)
}
\arguments{
\item{conn}{连接信息}

\item{app_id}{程序id}

\item{data}{用户数据}

\item{default_pwd}{默认密码}
}
\value{
返回值
}
\description{
上传用户数据
}
\examples{
userRight_upload()
}
