% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_baseUI.R
\name{var_action_button}
\alias{var_action_button}
\title{处理按纽控制的业务逻辑}
\usage{
var_action_button(id, init_null = FALSE)
}
\arguments{
\item{id}{内码}

\item{init_null}{初始化值是否为null}
}
\value{
返回值
}
\description{
处理按纽控制的业务逻辑
}
\examples{
var_action_button();
}
