% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\docType{methods}
\name{tsui,input.dateRange-method}
\alias{tsui,input.dateRange-method}
\title{Title}
\usage{
\S4method{tsui}{input.dateRange}(x)
}
\arguments{
\item{x}{input.dateRange.}
}
\value{
返回值
}
\description{
Title
}
\examples{
drawElement();
}
