% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_ui.R
\name{tsui,input.slider-method}
\alias{tsui,input.slider-method}
\title{2.2 定义tsui方法在input.slider类的实现}
\usage{
\S4method{tsui}{input.slider}(x)
}
\arguments{
\item{x}{input.slider. 输入对象实例}
}
\value{
返回值
}
\description{
2.2 定义tsui方法在input.slider类的实现
}
\examples{
tsui();
}
